Change method of checking for lethality of wounds
Just tally up all wounds across the body and compare to a global
threshold. Weights are applied to the wound damage based on the body
part, however these might not be necessary if we can find a suitable
threshold since the wound damage is already weighted based on the body
part.